requirements.txt
langchain==0.0.154
# langchain==0.0.137
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
altair==4.1.0
openai
tiktoken
